move getUser to helper folder
